Tried to debug ``EMDUnifrac_weighted_flow``
``diffab`` should be accumulated through ``for j in pos[i].union(neg[i])`` (in line 295), but it not was so.
To be clear, a variable ``diffab_amount`` was introduced instead of using ``diffab`` objects in different ways. 
I'd like to check if the diffab vectors for weighted and weighted_flow are same!